# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/workout/splits": {
      "post": {
        "description": "Create a new split. Any authenticated user can create splits. Official flag is automatically set to false.",
        "operationId": "SplitsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SplitRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Split created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "success": true,
                    "message": "Split created successfully",
                    "data": {
                      "id": "123e4567-e89b-12d3-a456-426614174000",
                      "name": "Upper/Lower",
                      "description": "4-day weekly program split into upper and lower.",
                      "routineIds": [
                        "123e4567-e89b-12d3-a456-426614174001"
                      ],
                      "trainingDays": [
                        "Monday",
                        "Tuesday",
                        "Thursday",
                        "Friday"
                      ],
                      "official": false,
                      "createdAt": "2024-01-01T12:00:00Z",
                      "updatedAt": "2024-01-01T12:00:00Z",
                      "routines": []
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Validation error. Invalid or missing required fields",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": [
                      "name must be longer than or equal to 3 characters",
                      "description must be longer than or equal to 10 characters",
                      "trainingDays must contain at least 1 elements",
                      "routineIds must contain at least 1 elements"
                    ],
                    "error": "Bad Request",
                    "statusCode": 400
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - No valid authentication token"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create split",
        "tags": [
          "Splits"
        ]
      },
      "get": {
        "description": "Get a paginated list of all splits. Accessible by all authenticated users. This action updates the user daily streak.",
        "operationId": "SplitsController_findAll",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of items per page",
            "schema": {
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "Number of items to skip",
            "schema": {
              "example": 0,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated split list retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "success": true,
                    "message": "Splits retrieved successfully",
                    "data": {
                      "items": [
                        {
                          "id": "123e4567-e89b-12d3-a456-426614174000",
                          "name": "Upper/Lower",
                          "description": "4-day weekly program split into upper and lower.",
                          "trainingDays": [
                            "Monday",
                            "Tuesday",
                            "Thursday",
                            "Friday"
                          ],
                          "official": false,
                          "createdAt": "2024-01-01T12:00:00Z",
                          "updatedAt": "2024-01-01T12:00:00Z",
                          "routines": []
                        }
                      ],
                      "total": 20,
                      "limit": 10,
                      "offset": 0
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - No valid authentication token"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List splits with pagination",
        "tags": [
          "Splits"
        ]
      }
    },
    "/workout/splits/{id}": {
      "get": {
        "description": "Retrieve a specific split by its ID. Accessible by all authenticated users.",
        "operationId": "SplitsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Split UUID",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Split retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "success": true,
                    "message": "Split retrieved successfully",
                    "data": {
                      "id": "123e4567-e89b-12d3-a456-426614174000",
                      "name": "Upper/Lower",
                      "description": "4-day weekly program split into upper and lower.",
                      "trainingDays": [
                        "Monday",
                        "Tuesday",
                        "Thursday",
                        "Friday"
                      ],
                      "official": false,
                      "createdAt": "2024-01-01T12:00:00Z",
                      "updatedAt": "2024-01-01T12:00:00Z",
                      "routines": [
                        {
                          "id": "123e4567-e89b-12d3-a456-426614174001",
                          "name": "Upper Body A"
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - No valid authentication token"
          },
          "404": {
            "description": "Split not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get split by ID",
        "tags": [
          "Splits"
        ]
      },
      "patch": {
        "description": "Update a split. Only the owner of the split or users with ADMIN role can update it.",
        "operationId": "SplitsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Split UUID",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SplitUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Split updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "success": true,
                    "message": "Split updated successfully",
                    "data": {
                      "id": "123e4567-e89b-12d3-a456-426614174000",
                      "name": "Upper/Lower",
                      "description": "4-day weekly program split into upper and lower.",
                      "trainingDays": [
                        "Monday",
                        "Tuesday",
                        "Thursday",
                        "Friday"
                      ],
                      "official": false,
                      "createdAt": "2024-01-01T12:00:00Z",
                      "updatedAt": "2024-01-01T12:00:00Z"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Validation error. Invalid field values",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": [
                      "name must be longer than or equal to 3 characters",
                      "trainingDays must contain at least 1 elements"
                    ],
                    "error": "Bad Request",
                    "statusCode": 400
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - No valid authentication token"
          },
          "403": {
            "description": "Forbidden - You can only modify splits that belong to you"
          },
          "404": {
            "description": "Split not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update split",
        "tags": [
          "Splits"
        ]
      },
      "delete": {
        "description": "Delete a split. Only the owner of the split or users with ADMIN role can delete it.",
        "operationId": "SplitsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Split UUID",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Split deleted successfully"
          },
          "401": {
            "description": "Unauthorized - No valid authentication token"
          },
          "403": {
            "description": "Forbidden - You can only delete splits that belong to you"
          },
          "404": {
            "description": "Split not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete split",
        "tags": [
          "Splits"
        ]
      }
    },
...
  },
  "info": {
    "title": "AthletIA API",
    "description": "API documentation for AthletIA",
    "version": "0.0.1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "SplitRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Split name",
            "minLength": 3,
            "maxLength": 50,
            "example": "Upper/Lower"
          },
          "description": {
            "type": "string",
            "description": "Split description",
            "minLength": 10,
            "maxLength": 500,
            "example": "4-day weekly program split into upper and lower."
          },
          "routineIds": {
            "description": "Included routine IDs",
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "trainingDays": {
            "type": "array",
            "description": "Training days - Valid values: Monday, Tuesday, Wednesday, Thursday, Friday, Saturday, Sunday",
            "items": {
              "type": "string",
              "enum": [
                "Monday",
                "Tuesday",
                "Wednesday",
                "Thursday",
                "Friday",
                "Saturday",
                "Sunday"
              ]
            }
          },
          "official": {
            "type": "boolean",
            "description": "Official split flag",
            "example": false
          }
        },
        "required": [
          "name",
          "description",
          "trainingDays",
          "official"
        ]
      },
      "SplitUpdate": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Split name",
            "minLength": 3,
            "maxLength": 50
          },
          "description": {
            "type": "string",
            "description": "Split description",
            "minLength": 10,
            "maxLength": 500
          },
          "routineIds": {
            "description": "Included routine IDs",
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "trainingDays": {
            "type": "array",
            "description": "Training days",
            "items": {
              "type": "string",
              "enum": [
                "Monday",
                "Tuesday",
                "Wednesday",
                "Thursday",
                "Friday",
                "Saturday",
                "Sunday"
              ]
            }
          },
          "official": {
            "type": "boolean",
            "description": "Official split flag",
            "example": false
          }
        }
      },
...
    }
  }
}